GE: Remove reduced number of conditional additions
In https://eprint.iacr.org/2023/059, we propose a trick (see 'Reducing the number of
conditional additions' paragraph) to reduce the number of conditional additions by
adding a fixed number of rows resulting in a very small chance (~2^-58) to wrongly
refusing a system of equations as not being solvable.
In the worst case, we wrongly reject a set of vinegars and sample new ones.
Unfortunately, this trick makes formal verification of the implementation much
harder (or impossible rather, because there is a small chance it is incorrect).
We, hence, disable it by default now.

See the corresponding change in the reference implementation:
https://github.com/pqov/pqov/pull/40